Enable and add further tests
In order to be able to mock your own methods (like waterfall), the
method you are testing must call the exact reference to the (to be
mocked) method that is also exported (see
http://stackoverflow.com/a/11598151/1316465 for further explenation).
So, I either had to call module.exports.METHOD or I could refactor the
code, making all methods properties of the main method. The first
option seemed weird and so I went with the latter.