Fix test bugs and make better use of sinon
There was a bug in the test for setRatingHistory titled 'Should call
main.waterfall once with an array of functions' where every call of func
was made with netflix.setVideoRating returning a rejected promise. func
was called synchronously and so every time it got to calling
netflix.setVideoRating, setVideoRating would be rejected. Due to another
bug described below, this did not cause the suite to fail but instead
only caused a weird log statement. This bug can be avoided by putting an
await before the expect().

There was another bug in some tests, where tests would be executed in a
forEach callback function. Those tests did not have any effect on the
whole suite failing (or the exit code) and so I replaced them with a
for-of loop.